refactor(desktop): organize worktree settings hooks
